+++ packages/messaging/src/events/post/payloads.ts
@@ -1,2 +1,12 @@
-export enum PostEventMessagingType {}
-// Empty for now
+export enum PostEventMessagingType {
+  POST_CREATED = 'post.created',
+  POST_DELETED = 'post.deleted',
+}
+
+export interface IPostCreatedPayload {
+  id: string;
+}
+
+export interface IPostDeletedPayload {
+  id: string;
+}