+++ packages/domain-event/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 3.4.4
+
+### Patch Changes
+
+- Updated dependencies [[`44e7a08`](https://github.com/Volontariapp/npm-packages/commit/44e7a0874947115ea18c2274c7f76d52df560ced)]:
+  - @volontariapp/messaging@2.7.4
+  - @volontariapp/shared@0.6.2
+  - @volontariapp/contracts@4.2.3
+  - @volontariapp/database@3.3.5
+  - @volontariapp/outbox@0.9.23
+
 ## 3.4.3
 
 ### Patch Changes