+++ packages/messaging/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 2.7.4
+
+### Patch Changes
+
+- post payload added
+
+- Updated dependencies [[`44e7a08`](https://github.com/Volontariapp/npm-packages/commit/44e7a0874947115ea18c2274c7f76d52df560ced)]:
+  - @volontariapp/shared@0.6.2
+  - @volontariapp/contracts@4.2.3
+
 ## 2.7.3
 
 ### Patch Changes